% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedData}
\alias{SimulatedData}
\title{Simulated Data
A simulated datalist for test and illustration. There are 100 subjects (50 cases and 50 controls) and 10 taxa.}
\format{
A data list with three components:
\describe{
\item{Treatment}{A treatment vector for 100 subjects. 0 represents the control group and 1 represents the case group.}
\item{otu.com}{A 100 x 10 numeric matrix containing compositional microbiome data. Each row is a subject, and each column is a
taxon. The row sum equals 1.}
\item{outcome}{A outcome vector for 100 subjects.}
}
}
\usage{
SimulatedData
}
\description{
Simulated Data
A simulated datalist for test and illustration. There are 100 subjects (50 cases and 50 controls) and 10 taxa.
}
\keyword{datasets}
